Rename words directory in models with word name
