+++ server/src/models/word.rs
